+++ src/ServerlessOffline.js
@@ -1,5 +1,6 @@
 import process, { exit } from "node:process"
-import { log } from "@serverless/utils/log.js"
+import { log, setLogUtils } from "./utils/log.js"
+import logSponsor from "./utils/logSponsor.js"
 import {
   commandOptions,
   CUSTOM_OPTION,
@@ -54,9 +55,10 @@ export default class ServerlessOffline {
     "offline:start:ready": this.#ready.bind(this),
   }
 
-  constructor(serverless, cliOptions) {
+  constructor(serverless, cliOptions, utils) {
     this.#cliOptions = cliOptions
     this.#serverless = serverless
+    setLogUtils(utils)
   }
 
   // Entry point for the plugin (sls offline) when running 'sls offline start'
@@ -64,6 +66,12 @@ export default class ServerlessOffline {
     this.#mergeOptions()
 
     this.#preLoadModules()
+
+    if (this.#cliOptions.noSponsor) {
+      log.notice()
+    } else {
+      logSponsor()
+    }
 
     const {
       albEvents,
@@ -279,13 +287,11 @@ export default class ServerlessOffline {
       origin: this.#options.corsAllowOrigin,
     }
 
-    log.notice()
     log.notice(
       `Starting Offline at stage ${
         this.#options.stage || provider.stage
       } ${gray(`(${this.#options.region || provider.region})`)}`,
     )
-    log.notice()
     log.debug("options:", this.#options)
   }
 
@@ -420,13 +426,13 @@ export default class ServerlessOffline {
   }
 
   #preLoadModules() {
-    const { preLoadModules } = this.#options
-    const modules = preLoadModules.split(",")
+    const modules = this.#options.preLoadModules.split(",")
+
     modules.forEach((module) => {
       try {
         import(module)
-      } catch (err) {
-        log.error(`Error importing module ${module}: ${err}`)
+      } catch (error) {
+        log.error(`Error importing module ${module}: ${error}`)
       }
     })
   }